Giant refactoring (#28)
* telegram types

* fixing issues

* fixing issues

* fixing telegram model

* firebase model

* fixing model

* code review

* refactoring

* refactoring

* command design

* code review

* new model

* refactoring

* new arhitecture

* fixing design

* new flow

* user polish

* code review

* comments

* started reworking handlers

* refactoring

* language request & setup

* last note refactoring

* fixing issues

* comment out old code

* fixing runtime crashes

* fixing memory stuff

* fixing memory issues

* clean up

* fixing memory issues

* fixing issues

* clean up

* temporary fix

* onboarding

* first user fix

* clean up

* modified onboarding

* clean up

* firestore clean up

* timezone tests

* no exising translation id test

* comments

* comments

* comments

* more tests

* job testing

* more tests

* comments

* user tests

---------

+++ internal/telegram/telegram.go
@@ -15,6 +15,7 @@ func init() {
 	baseURL = "https://api.telegram.org/bot" + os.Getenv("CAPY_TELEGRAM_BOT_TOKEN")
 }
 
+// Parse an incoming update
 func Parse(r *http.Request) *Update {
 	var update Update
 	if err := json.NewDecoder(r.Body).Decode(&update); err != nil {
@@ -24,7 +25,8 @@ func Parse(r *http.Request) *Update {
 	return &update
 }
 
-func SendMessage(chatID int, text string, replyMarkup *InlineKeyboardMarkup) {
+// Send a message to a chat
+func SendMessage(chatID int64, text string, replyMarkup *InlineKeyboardMarkup) {
 	var url string = baseURL + "/sendMessage"
 
 	message := SendMessageRequest{
